Add notificacao AcaoCritica, Add salvar e atualizar AcaoCritica em Lote

+++ src/main/java/br/com/tellescom/service/AcaoCriticaService.java
@@ -1,16 +1,15 @@
 package br.com.tellescom.service;
 
 import br.com.tellescom.domain.AcaoCritica;
+import br.com.tellescom.domain.enumeration.EnumTipoCritica;
 import br.com.tellescom.repository.AcaoCriticaRepository;
 import br.com.tellescom.service.dto.AcaoCriticaDTO;
 import br.com.tellescom.service.mapper.AcaoCriticaMapper;
 
 import java.time.ZonedDateTime;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
+import java.util.*;
 import java.util.stream.Collectors;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
@@ -29,9 +28,12 @@ public class AcaoCriticaService {
 
     private final AcaoCriticaMapper acaoCriticaMapper;
 
-    public AcaoCriticaService(AcaoCriticaRepository acaoCriticaRepository, AcaoCriticaMapper acaoCriticaMapper) {
+    private final NotificacaoUsuarioService notificacaoUsuarioService;
+
+    public AcaoCriticaService(AcaoCriticaRepository acaoCriticaRepository, AcaoCriticaMapper acaoCriticaMapper, NotificacaoUsuarioService notificacaoUsuarioService) {
         this.acaoCriticaRepository = acaoCriticaRepository;
         this.acaoCriticaMapper = acaoCriticaMapper;
+        this.notificacaoUsuarioService = notificacaoUsuarioService;
     }
 
     /**
@@ -44,8 +46,11 @@ public class AcaoCriticaService {
         LOG.debug("Request to save AcaoCritica : {}", acaoCriticaDTO);
         AcaoCritica acaoCritica = acaoCriticaMapper.toEntity(acaoCriticaDTO);
         acaoCritica.setCriadoEm(ZonedDateTime.now());
+        acaoCritica.setIsNotificado(true);
         acaoCritica = acaoCriticaRepository.save(acaoCritica);
-        return acaoCriticaMapper.toDto(acaoCritica);
+        AcaoCriticaDTO result = acaoCriticaMapper.toDto(acaoCritica);
+        notificacaoUsuarioService.enviarEmailAcaoCritica(EnumTipoCritica.CRIAR, result);
+        return result;
     }
 
     /**
@@ -59,7 +64,9 @@ public class AcaoCriticaService {
         AcaoCritica acaoCritica = acaoCriticaMapper.toEntity(acaoCriticaDTO);
         acaoCritica.setAtualizadoEm(ZonedDateTime.now());
         acaoCritica = acaoCriticaRepository.save(acaoCritica);
-        return acaoCriticaMapper.toDto(acaoCritica);
+        AcaoCriticaDTO result = acaoCriticaMapper.toDto(acaoCritica);
+        notificacaoUsuarioService.enviarEmailAcaoCritica(EnumTipoCritica.ATUALIZAR, result);
+        return result;
     }
 
     /**
@@ -110,16 +117,46 @@ public class AcaoCriticaService {
      *
      * @param id the id of the entity.
      */
+    @Transactional
     public void delete(Long id) {
         LOG.debug("Request to delete AcaoCritica : {}", id);
-        acaoCriticaRepository.deleteById(id);
+        AcaoCritica acaoCritica = acaoCriticaRepository.findById(id).orElse(null);
+        if (acaoCritica != null) {
+            AcaoCriticaDTO acaoCriticaDTO = acaoCriticaMapper.toDto(acaoCritica);
+            notificacaoUsuarioService.enviarEmailAcaoCritica(EnumTipoCritica.CANCELAR, acaoCriticaDTO);
+            acaoCriticaRepository.deleteById(id);
+        }
     }
 
-    public Set<AcaoCriticaDTO> findAllByIndicadorCritica(Long idIndicadorCritica){
+    public Set<AcaoCriticaDTO> findAllByIndicadorCritica(Long idIndicadorCritica) {
         LOG.debug("Request to return a SET of AcaoCriticaDTO from Id of IndicadorMeta: {}", idIndicadorCritica);
         return acaoCriticaRepository.findAllByIndicadorCritica(idIndicadorCritica)
             .stream()
             .map(acaoCriticaMapper::toDto)
             .collect(Collectors.toSet());
+    }
+
+    public List<AcaoCriticaDTO> salvaAcaoCriticaEmLote(List<AcaoCriticaDTO> request) {
+        LOG.debug("Request to save a lot of AcaoCriticaDTO: {}", request);
+        return request.stream().map(this::save).toList();
+    }
+
+
+    public List<AcaoCriticaDTO> atualizaAcaoCriticaEmLote(List<AcaoCriticaDTO> request) {
+        LOG.debug("Request to update a lot of AcaoCriticaDTO: {}", request);
+        return request.stream()
+            .map(item -> {
+                try {
+                    if (item.getId() == null) {
+                        return this.save(item);
+                    }
+                    return this.update(item);
+                } catch (Exception e) {
+                    LOG.error("Error processing item: {}", item, e);
+                    return null;
+                }
+            })
+            .filter(Objects::nonNull)
+            .toList();
     }
 }
